% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_label.R
\name{add_label}
\alias{add_label}
\title{Consistent labelling for base plots}
\usage{
add_label(xfrac, yfrac, label, pos = 4, ...)
}
\arguments{
\item{xfrac}{is label position, in fraction from x axis}

\item{yfrac}{is label position, in fraction from y axis}

\item{label}{is text to add}

\item{pos}{is text alignment (see ?text)}
}
\description{
For adding easy labels to multipanel figures (a,b,c). Stolen from Sean Anderson (https://seananderson.ca/2013/10/21/panel-letters/)
}
\examples{
add_label
}
\keyword{plot}
