% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_dfa.R
\name{fit_DFA}
\alias{fit_DFA}
\title{Fit DFA to timeseries datasets}
\usage{
fit_DFA(df, proc = 3, id, plot = TRUE, yr_frst, yrlast, clr = clr)
}
\arguments{
\item{df}{is a matrix of time series}

\item{proc}{is the number of underlying processes, defaults to 3}

\item{id}{is the name of each time series}

\item{plot}{specifies whether fits and loadings should be plotted, defaults to TRUE}

\item{clr}{is a vector of colours applied to each time series if plot = TRUE}

\item{yr_first}{is the first year of the time series}

\item{yr_last}{is the last year of the time series}
}
\description{
This builds on DFA tutorial here: https://nwfsc-timeseries.github.io/AFTSLabbook/sec-dfa-plot-data.html
}
\examples{
fit_DFA
}
\keyword{DFA}
