% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_var_importance.R
\name{gam_dev_exp}
\alias{gam_dev_exp}
\title{Variable importance for GAMs}
\usage{
gam_dev_exp(
  dataset,
  exp.names,
  indicator,
  family,
  base_k = -1,
  smoother = "cr"
)
}
\arguments{
\item{dataset}{is the dataset containing y and all x covariates. should be scaled and centered (mean = 0, sd = 1)}

\item{exp.names}{= explanatory covariate names, passed as vector of characters}

\item{indicator}{= y variable of interest (character)}

\item{family}{= GAM family distribution, takes 'gaussian' or 'Gamma'}

\item{base_k}{= number of knots for each smoother term. Default = -1, mgcv  uses generalized cross-validation}

\item{smoother}{= cubic regression spline method (cr is default)}
}
\description{
Function builds GAM and calculates deviance explained by each covariate. 
Simon Wood; http://r.789695.n4.nabble.com/variance-explained-by-each-term-in-a-GAM-td836513.html
Litzow et al. (2013) Reassessing regime shifts in the North Pacific: incremental climate change and commercial fishing are necessary for explaining decadal-scale biological variability. Global Change Biology.
}
\examples{
gam.dev.exp
}
\keyword{gam}
