% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_jacknife.R
\name{gam_jacknife}
\alias{gam_jacknife}
\title{Fit jacknife GAM analysis.}
\usage{
gam_jacknife(
  dataset,
  exp.names,
  indicator,
  family,
  base_k = -1,
  smoother = "cr",
  n.param.max = 3
)
}
\arguments{
\item{dataset}{is a data frame of response and explanatory covariates}

\item{exp.names}{is a vector of explanatory covariate names, matching colnames of dataset}

\item{indicator}{is the name of the focal response variable}

\item{family}{specifies the fitted distribution (e.g. 'gaussian', 'poisson', 'binomial')}

\item{base_k}{is the number of knots for each smoother term. Defaults to -1, where mgcv uses generalized cross-validation}

\item{smoother}{is the spline basis, defaults to 'cr'}

\item{n.param.max}{is the maximum number of fitted explanatory covariates, defaults to 3}
}
\description{
This function runs a drop-one jacknife to estimate relative variable importance.
}
\examples{
gam.jacknife
}
\keyword{gam}
