% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_predict.R
\name{gam_predict}
\alias{gam_predict}
\title{GAM predictions}
\usage{
gam_predict(dataset, exp.names, indicator, family, base_k = -1,
  smoother = "cr", n.param.max = 3)
}
\arguments{
\item{dataset}{= dataset containing y and all x covariates. should be scaled and centered (mean = 0, sd = 1)}

\item{exp.names}{= explanatory covariate names, passed as vector of characters}

\item{indicator}{= y variable of interest (character)}

\item{family}{= GLM family distribution, takes 'gaussian' or 'Gamma'}

\item{base_k}{= number of knots for each smoother term. Default = -1, mgcv  uses generalized cross-validation}

\item{smoother}{= cubic regression spline method (cr is default)}

\item{n.param.max}{= maximum number of parameters included in each reduced model}
}
\description{
Function builds GAM and dredges saturated models, fixed to maximum of 3 covariates to prevent overfitting, then predicts response against range of predictor covariates, for either 
1. the top-ranked model (AICc > 2 lower than candidate models) OR 2. across a top-ranking model set (AICc < 7 of top-ranked model)
}
\examples{
gam.predict
}
\keyword{multimodel}
