% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_predict_newdata.R
\name{gam_predict_newdata}
\alias{gam_predict_newdata}
\title{Testing predictive power for GAMs}
\usage{
gam_predict_newdata(
  dataset,
  newdataset,
  exp.names,
  indicator,
  family,
  base_k = -1,
  smoother = "cr",
  n.param.max = 3
)
}
\arguments{
\item{dataset}{= dataset containing y and all x covariates. should be scaled and centered (mean = 0, sd = 1)}

\item{newdataset}{= generating predictions with 'newdata'. should be scaled + centered (mean = 0, sd = 1)}

\item{exp.names}{= explanatory covariate names, passed as vector of characters}

\item{indicator}{= y variable of interest (character)}

\item{family}{= GAM family distribution, takes 'gaussian' or 'Gamma'}

\item{base_k}{= number of knots for each smoother term. Default = -1, mgcv  uses generalized cross-validation}

\item{smoother}{= cubic regression spline method (cr is default)}

\item{n.param.max}{= maximum number of parameters included in each reduced model}
}
\description{
Function builds GAM and dredges saturated models, fixed to maximum of 3 covariates to prevent overfitting. Then predicts response against new predictor covariates
1. the top-ranked model (AICc > 2 lower than candidate models) OR
2. across a top-ranking model set (AICc < 7 of top-ranked model)
}
\examples{
gam.predict.newdata
}
\keyword{GAM,}
\keyword{multimodel,}
\keyword{prediction}
