% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfa_fits.R
\name{get_DFA_fits}
\alias{get_DFA_fits}
\title{Examine fits from a DFA object}
\usage{
get_DFA_fits(MLEobj, dd = NULL, alpha = 0.05)
}
\arguments{
\item{MLEobj}{is fitted MARSS DFA object}

\item{dd}{is explanatory covariates}

\item{alpha}{is significance level}
}
\description{
For DFA fitted to time series datasets. This builds on DFA tutorial here: https://nwfsc-timeseries.github.io/AFTSLabbook/sec-dfa-plot-data.html
}
\examples{
get_DFA_fits
}
\keyword{DFA}
