% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMI_tvalues.R
\name{glm.avg.tvalue}
\alias{glm.avg.tvalue}
\title{Multimodel averaging with t-values}
\usage{
glm.avg.tvalue(dataset, exp.names, indicator, family)
}
\arguments{
\item{dataset}{= dataset containing y and all x covariates. should be scaled and centered (mean = 0, sd = 1)}

\item{exp.names}{= explanatory covariate names, passed as vector of characters}

\item{indicator}{= y variable of interest (character)}

\item{family}{= GLM family distribution, takes 'gaussian' or 'Gamma'}
}
\description{
Function builds GLM, runs multi-model inference to produce: 
1. Standardized t-values (and standard error) for each covariate across full model set (effect sizes)
2. R-squareds of each model < 7 AICc of top model
3. Predicted y across each covariate, weighted by AIC for all models < 7 AICc of top model (and variance in predictions)
}
\examples{
glm.avg.tvalue
}
\keyword{multimodel}
