% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_gam_test.R
\name{glm.gam.test}
\alias{glm.gam.test}
\title{Comparison of GAM and GLM models}
\usage{
glm.gam.test(dataset, exp.names, indicator, family)
}
\arguments{
\item{dataset}{= dataset containing y and all x covariates. should be scaled and centered (mean = 0, sd = 1)}

\item{exp.names}{= explanatory covariate names, passed as vector of characters}

\item{indicator}{= y variable of interest (character)}

\item{family}{= GLM family distribution, takes 'gaussian' or 'Gamma'}
}
\description{
Function builds GLM or GAM, compares AIC and outputs the winner.
}
\examples{
glm.gam.test
}
\keyword{comparison}
\keyword{model}
