% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairs2.R
\name{pairs2}
\alias{pairs2}
\title{Pairs plots for correlations}
\usage{
pairs2(x, labels, panel = points, ..., lower.panel = panel.smooth2,
  upper.panel = panel.cor, diag.panel = panel.hist,
  text.panel = textPanel, label.pos = 0.5 + has.diag/3,
  cex.labels = NULL, font.labels = 2, row1attop = TRUE, gap = 1)
}
\arguments{
\item{x}{is dataset containing y and all x covariates.}
}
\description{
pairs2 creates Pearson correlations and scatterplots with LOESS for data diagnostics.
}
\examples{
pairs2
}
\keyword{multimodel}
