% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-cor-functions.R
\name{panel.cor}
\alias{panel.cor}
\title{Pairs plots for correlations}
\usage{
panel.cor(x, y, digits = 2, prefix = "", cex.cor)
}
\arguments{
\item{x}{is dataset containing y and all x covariates.}
}
\description{
pairs2 creates Pearson correlations and scatterplots with LOESS for data diagnostics.
}
\examples{
pairs2
}
\keyword{multimodel}
