% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_hist.R
\name{panel.hist}
\alias{panel.hist}
\title{Pairs plots for correlations}
\usage{
panel.hist(x, ...)
}
\arguments{
\item{x}{dataset containing y covariate}
}
\description{
panel.hist creates histograms for pairs2
}
\examples{
panel.hist
}
\keyword{multimodel}
