% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_smooth2.R
\name{panel.smooth2}
\alias{panel.smooth2}
\title{Pairs plots for correlations}
\usage{
panel.smooth2(
  x,
  y,
  col = par("col"),
  bg = alpha("grey", 0.5),
  pch = 21,
  cex = 1,
  span = 2/3,
  iter = 3,
  ...
)
}
\arguments{
\item{x}{is x covariate}

\item{y}{is y covariate}
}
\description{
panel.smooth2 creates scatterplot with LOESS for data diagnostics.
}
\examples{
panel.smooth2
}
\keyword{multimodel}
