% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling_function.R
\name{scaler}
\alias{scaler}
\title{Scaling and centering explanatory covariates}
\usage{
scaler(df, ID, centered = TRUE, scaled = TRUE, ...)
}
\arguments{
\item{df}{= dataset containing all response, explanatory, and grouping factors}

\item{ID}{= names of grouping variables that should not be scaled; provide as c("Var1", "Var2")}

\item{centered}{= should the function center the mean at 0? (defaults to TRUE)}

\item{scaled}{= should the function scale each variable by its standard deviation? (defaults to TRUE)}

\item{family}{= GLM family distribution, takes 'gaussian' or 'Gamma'}
}
\description{
function to scale numeric and categorical variables for multi-model approaches
}
\examples{
scaler
}
\keyword{demean,}
\keyword{multimodel,}
\keyword{scale}
