% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling_function.R
\name{scaler_cat}
\alias{scaler_cat}
\title{Scaling and centering explanatory covariates}
\usage{
scaler_cat(df, ID, centered = TRUE, scaled = TRUE, cats = TRUE, ...)
}
\arguments{
\item{df}{= dataset containing all response, explanatory, and grouping factors}

\item{ID}{= names of grouping variables that should not be scaled; provide as c("Var1", "Var2")}

\item{centered}{= should the function center the mean at 0? (defaults to TRUE)}

\item{scaled}{= should the function scale each variable by its standard deviation? (defaults to TRUE)}

\item{cats}{= should the function convert categorical variables into dummies?}
}
\description{
function to scale numeric and categorical variables for multi-model approaches. Categorical covariates must be provided as factors and will be converted to 0,1 dummy variables.
}
\examples{
scaler_cat
}
\keyword{demean,}
\keyword{multimodel,}
\keyword{scale}
