% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themeSleek.R
\name{theme_sleek}
\alias{theme_sleek}
\title{A gg theme Function}
\usage{
theme_sleek(base_size = 11, base_family = "", position = "standard")
}
\description{
This function is for publication quality ggplots. It is a copy of Sean Anderson's ggsidekick which is no longer supported, and Cameron Freshwater added argument for top, bottom, middle for multipanel functionality.
}
\examples{
theme_sleek()
}
\keyword{ggplot}
