% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vifmer.R
\name{vif.mer}
\alias{vif.mer}
\title{Multimodel averaging with t-values}
\usage{
vif.mer(fit)
}
\arguments{
\item{fit}{= fitted model}
}
\description{
variance inflation factors for lme4 package: https://github.com/aufrank/R-hacks/blob/master/mer-utils.R. Adapted from rms::vif
}
\examples{
vif.mer
}
